requirements.txt
flask
moviepy
openai-whisper
